Version 3.3
Anton Sharnov thinks it is better to leave the foreground color as is for spell error in 256-color cterm. If you agree this, the version 3.3 is for you, else you can just edit the source. Also, your can search for text 'transparency' within the peaksea.vim to know how to get cterm background transparency.